backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
firebase-admin==6.7.0
pydantic==2.11.3
google-cloud-secret-manager==2.18.2
python-multipart==0.0.9
aiofiles==23.2.1
numpy==1.26.4
requests==2.31.0
aiohttp==3.9.3
qdrant-client==1.7.3
sentence-transformers==2.5.1
llama-index==0.10.5
torch==2.6.0
transformers==4.38.2 